ARM: dts: sun7i: Add dts file for the A20-linova1-7 HMI for debugging
Since LiNova1 doesn't have an ethernet interface and USB0 is a Usb host,
force it to work as an Usb Device port.
This way RNDIS can be enable over it and use it as a network card.
Then you can use remote gdb.

Add overlay dts with Usb0 marked as "otg" instead of "host".

Signed-off-by: Giulio Benetti <giulio.benetti@micronovasrl.com>